\name{EigenImpute}
\alias{EigenImpute}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Gene Expression Data Enhancer: Improving the Quality of Gene Expression Data using an Internal Prediction Strategy}
\description{
  TBD. 
}
\usage{
EigenImpute(EstObj, Ymiss)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{EstObj}{ An object produced by RobEst() which contains
    information about the covariance matrix of the data. }
  \item{Ymiss}{ An nxp-dimensional data matrix with missing values to be
    imputed.}
}
\details{
  TBD.
}
\value{ An nxp-dimensional data matrix will all missing values imputed. }
\references{ TBD. }
\author{Xing Qiu}
\seealso{  }
\examples{
# to be written later.
}
