\name{GEDE-package}
\alias{GEDE-package}
\alias{GEDE}
\docType{package}
\title{
  Gene Expression Data Enhancer: Improving the Quality of Gene Expression Data using an Internal Prediction Strategy.
}
\description{
  This package includes ....
}
\details{
\tabular{ll}{
Package: \tab GEDE\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2023-04-13\cr
License: \tab GPL 2.0\cr
LazyLoad: \tab yes\cr
}
Please see the help files of individual functions.
}
\author{
  Xing Qiu
Maintainer: Xing Qiu <xing.qiu@gmail.com>
}
\references{
  Some text book.
}
\keyword{package}
\examples{
% X <- randomization(50, ngroups=2, nblocks=5)
% randomization.latex(X)
}
